Conflicts:
	runtimes/cpp/platforms/android/MoSyncBridge.cpp

+++ runtimes/cpp/base/WaveAudioSource.cpp
@@ -19,6 +19,8 @@ Software Foundation, 59 Temple Place - Suite 330, Boston, MA
 #include "WaveAudioSource.h"
 #include "Stream.h"
 #include "StreamHelpers.h"
+#include <stdint.h>
+
 using namespace Base;
 using namespace MoSyncError;
 
@@ -108,7 +110,7 @@ int WaveAudioSource::init()
 {
 	struct chunk_hdr {
 		char  id[4];
-		unsigned long len;
+		uint32_t len;
 	};
 
 	chunk_hdr chnk;
@@ -131,8 +133,8 @@ int WaveAudioSource::init()
 			struct fmt_header {
 				short          wFormatTag;
 				unsigned short wChannels;
-				unsigned long  dwSamplesPerSec;
-				unsigned long  dwAvgBytesPerSec;
+				uint32_t  dwSamplesPerSec;
+				uint32_t  dwAvgBytesPerSec;
 				unsigned short wBlockAlign;
 				unsigned short wBitsPerSample;
 			};
