+++ runtimes/cpp/platforms/iphone/Classes/UI/Widgets/ButtonWidget.mm
@@ -49,7 +49,7 @@
 	MAWidgetEventData *eventData = new MAWidgetEventData;
 	eventData->eventType = MAW_EVENT_CLICKED;
 	eventData->widgetHandle = handle;
-	event.data = eventData;
+	event.data = (int)eventData;
 	Base::gEventQueue.put(event);
 }
