+++ runtimes/cpp/platforms/iphone/Classes/UI/Widgets/ListViewWidget.mm
@@ -54,7 +54,7 @@
 	eventData->eventType = MAW_EVENT_ITEM_CLICKED;
 	eventData->widgetHandle = handle;
 	eventData->listItemIndex = index;
-	event.data = eventData;
+	event.data = (int)eventData;
 	Base::gEventQueue.put(event);	
 }
